test: Added LaTeXTextBox.test.ts as valid Jest tests
Fixed name of LaTeXTextBox.test.ts so that it would be recognized by Jest as a file containing tests.
Fixed definitions of incompleteLaTeXTextBox2 and completeLaTeXTextBox in LaTeXTextBox.test.ts to use the correct name latex_text instead of link for its string member, so that all tests in that test suite now pass.